experimentations/llama4/requirements.txt

#for example:
#torch==2.0.1
#matplotlib==3.7.2
llama_api_client==0.1.1
transformers==4.52.4
synthetic-data-kit==0.0.4b1
llama-prompt-ops==0.0.7
openai==1.82.1
together==1.5.8
pydantic==2.11.5
pdf2image==1.17.0
pypdf==5.5.0
opencv-python==4.11.0.86
torchvision==0.22.0
datasets==3.6.0
python-dotenv==1.0.1


sentence-transformers==2.7.0
langchain
langchain_community